+++ libs.py
@@ -1,0 +1,7 @@
+
+def read_file(path):
+  content = ''
+  # ignore emoji
+  with open(path, encoding='utf-8', errors='ignore') as f:
+    content = f.readlines()
+  return content
